manual: Describe the GNU Boot project.
So far the manual only tell that GNU Boot is a boot software
distribution and it explains what it means.

It didn't tell what it means for GNU Boot to be fully free.

In addition, other 100% free software boot distributions also exists,
so we also need to explain why we need GNU Boot to exist.

More details about the GPU issue will be added later on.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>