website: status: add 0.1 RC4 with the 4 image types.
Before we had columns for installation and upgrade. These got replaced
to add the 4 image types as space was missing for all of them: it
would have made the markdown lines over 80 characters and it may also
create tables whose width is too big to be readable.

The installation and upgrade instructions were not added in an
additional table because:

- so far nobody tested trying to find instructions on the GNU Boot
  website to install GNU Boot,

- we plan to (re)write install and upgrade instructions to make them
  easier to use than what is currently scattered around the website.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>