resources: grub: config: fix wrong dev/device variable.
This was not caught because the dev variable is used earlier and so it
produces no error.

In addition the commit 860b00bf1e798d86c8bb2a70d77633599dfa1da2
("ressources/grub, website: add LVM2 support (fix bug #65663 "No
support for LVM2").") where this issue was introduced was not tested
with LVM partitions that are encrypted.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>