packages: roms: tests: Fix Trisquel LVM test.
This was broken from the start in the commit
9cc02ddde1e164fabfbddc8bbd3832ef9468d92d ("packages: roms: Start
adding automatic tests.").

I vaguely remember having made the original tests with Trisquel 10,
and if we select LVM in Trisquel 10 both in the graphical and
netinstall installers, it ends up creating a BIOS boot partition (code
ef02), an UEFI partition (code ef00) in the case of the graphical
installer, and an LVM physical volume. I then migrated the automatic
test to Trisquel 11 before sending the patch.

But with the commit 9cc02ddde1e164fabfbddc8bbd3832ef9468d92d
("packages: roms: Start adding automatic tests."), we end up with a
rootfs that has a /boot partition, so GRUB will find that and use it
to boot, without testing if GRUB can find the grub.cfg inside the LVM
partition.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>