website: status: tables: workaround lack of colspan.
When using ./configure --disable-guix, and a recent distribution (like
Parabola and Guix but not Trisquel 11 (aramo)), we can see cells
spanning multiple columns.

Once the website is built (with make build), we can easily test for
that with grep:
    $ grep colspan ./untitled/www/lbwww/site/status.html
    <td colspan="4">Can’t install due to missing images</td>
    <td colspan="4">Fail: #66487</td>
    <td colspan="4">Untested | Untested | Untested | Untested</td>
    <td colspan="4">Can’t install due to missing images</td>
    <td colspan="4">Fail: #66463, #66487</td>
    <td colspan="4">Untested | Untested | Fail: #66487 | Fail: #66487</td>
    <td colspan="3">Can’t install due to missing images</td>
    <td colspan="3">Can’t install due to missing images</td>
    <td colspan="3">Can’t install due to missing images</td>

But with ./configure without --disable-guix, we don't see any columns
spanning since the commit 338154a65451c088f9d14334e64e2bb12db255d7
("website: Makefile.am: use common guix shell command.").

If we build the website at this commit (without --disable-guix) we see
that the '--commit=' argument is bogus in the command being used:
    guix time-machine --commit= -- shell --system=i686-linux
    --container --network --emulate-fhs --share=`realpath ../` bash
    coreutils findutils git grep nss-certs pandoc sed -- ./build.sh

And we then ends up with this message:
    guix time-machine: error: Git error: unable to parse OID - too short

This was then repaired in the commit
2e3818668d1d65ae02c508851885a7a812ec23cb ("website: use Guix
v1.4.0."), but we then end up without 'colspan' because we ended up
using older software that don't support this feature.

There are two reasons for not relying yet on this feature:
- We want to use Guix v1.4.0.
- We build a tarball of the website is also part of the release, and
  guix is not used during the build. And we need all distributions to
  produce at the very least a functionally equivalent output.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>