release: tests: add debug images to the list of expected files.
Without that fix, make release fails with the following error:
    [ !! ] release/roms-dbg/gnuboot-<version>_debug_x200_8mb.tar.xz
           missing in ${release_files}
    [ !! ] release/roms-dbg/gnuboot-<version>_debug_t400_8mb.tar.xz
           missing in ${release_files}
    [ !! ] release/roms-dbg/gnuboot-<version>_debug_kgpe-d16-rdimm_2mb.tar.xz
           missing in ${release_files}
    [ !! ] release/roms-dbg/gnuboot-<version>_debug_qemu-pc_2mb.tar.xz
           missing in ${release_files}
with <version> being the GNU Boot version being built.

Also note that this was a known issue when it was introduced in the
commit b4a5a47499d5bf0cfe9c1fa137b96845de08450c ("packages: roms:
release: include debug images in the release") and I was aware of it
during the review.

The issue was not fixed at the time because we had planned a 100% free
software and/or FSDG install party where we installed GNU Boot.

And for that install party I wanted to make sure people were using a
tagged release that had the commit
860b00bf1e798d86c8bb2a70d77633599dfa1da2 ("ressources/grub, website:
add LVM2 support (fix bug #65663 "No support for LVM2").").

This commit was supposed to support some configurations where there is
no /boot and the rootfs is on some LVM2 partition (but in reality this
commit contained issues and so this got fixed later on).

Since we needed to make a release of GNU Boot just before the install
party and to allow time to test the images as well (especially because
this is done by volunteers), I pushed for cutting every corner
possible to make sure that the images were released in time and to
allow time for testing. And so I argued for not fixing the tests for
the images.

At the end of the day the images were released in time but we had 0
time for doing tests (they were done later on).

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>