website: status: KCMA-D8: split status in ECC/non-ECC.
There is currently different images for use with ECC or non-ECC RAM:
the rdimm images are for use with ECC memory and the udimm are for use
with non-ECC memory.

Thanks a lot to isf on #gnuboot on libera chat for the correct Spanish
translation of non-ECC (before isf's fix, I added a space instead of a
hyphen because Non-binary gender is translated Género no binario
according to Spanish and English Wikipedia).

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>