website: global.css: fix titles top and down margins
Without this fix, if there is a table and a heading right after it,
the heading will be stuck to the bottom of the table.

This is for instance visible in "GNU Boot 0.1 RC3 status" in
website/site/software/gnuboot/web/status.html where the "Stability:"
h2 is stuck to the bottom of the GNU Boot 0.1 RC3 status table.

This commit corrects this by using simple margins.

Signed-off-by: Adrien 'neox' Bourmault <neox@gnu.org>
GNUtoo: improved commit message.
Acked-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>