website: docs: update main page translation.
While the Spanish translation of the main page made it into the GNU
Boot git after the commit 413308f7a7164fc7b497c5fe4babdd67843f152f
("site: Add project status on the main page and move contribution
information."), for some reasons it translated a version of the main
page before this change.

Updating both the main page and translating the contribute page as
well enable to bring the Spanish main page back in sync with the
English one (because both pages were modified by the commit mentioned
above).

This is especially important because there is some effort in GNU Boot
to keep the translation in sync with the English version, especially
on the GNU Boot status page, and because of that any drift between the
English and Spanish version of a page makes it even more difficult for
non-native speakers to keep everything in sync.

GNUtoo: commit message, tiny markdown fix, whitespace and line
        breaks fixes, restore spanish links.
Acked-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>