website: docs: contribute: fix markdown link syntax.
Without that fix, on the deployed GNU Boot website, the Bug-gnuboot
link appears as "[Bug-gnuboot]
(https://lists.gnu.org/mailman/listinfo/bug-gnuboot)" inside the HTML
instead of appearing as a link (<a href="[...]">Bug-gnuboot</a>), and
the link in "Technical contributions" has the same issue.

The Savannah link was also made clickable along the way.

This was broken from the start by the commit
413308f7a7164fc7b497c5fe4babdd67843f152f ("site: Add project status on
the main page and move contribution information.").

Signed-off-by: Alex <alex@puer-robustus.eu>
GNUtoo: add "Without that fix [...]" and "The Savannah link [...]" in
        the commit message and mentioned the broken commit.
Acked-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>