manual: rework the introduction.
The introduction had several issues, they are related to the term
"memory chip", but also to the way it is explained:

- The manual was read by 3 less technical people, and one of them
  understood that GNU Boot was to be installed on the RAM, and the
  other two didn't understand where GNU Boot was supposed to be
  installed. Here it is pretty clear that "memory chip" looks related
  to the RAM and is therefor confusing.

- The term "memory chip" is usually not used to refer to the flash
  chip in projects like Coreboot or other boot firmware distributions,
  and in Coreboot it even refers to the RAM chips: from the
  Documentation/acronyms file: "RAMID - Boards that have soldered-down
  memory (no DIMMs) can have various different sizes, speeds, and
  brands of memory chips attached.". The consequence is that, if we
  kept "memory chip", contributors to GNU Boot would need to acquire
  GNU Boot specific knowledge instead of being able to reuse knowledge
  of the field GNU Boot operates in.

Because of all that the term was changed to "boot flash" that is
sometime used in the field, and also because it is extremely precise
(computers have a lot of flash chip everywhere for many purposes), and
the introduction was reworked as well in the hope that it would make
things easier to understand for less technical people.

Thanks to Adrien 'neox' Bourmault for the feedback on the issue, and
to the people that were in contact with Adrien and that read the
manual.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>