manual: improve the description of Canoeboot.
The "Because of that it can be harder for users to use." is not
clear. My assumption was that features and computers came and go, but
I didn't manage to confirm that and this slept through the review
process.

What seems to have happens instead was that some things are dropped
and don't seem to come back. An example of that is the build of
installation utilities for i945 ThinkPads.

However things changed a lot since these utilities were dropped, and
so after a lengthy discussion with Leah Rowe, together we managed to
nail down more practical differences between GNU Boot and Canoeboot
that nail down to different approaches, which are, at the end of the
day complementary.

Thanks a lot to Leah Rowe for all the help during the review of this
patch.

Reported-by: Leah Rowe <leah@libreboot.org>
Link: https://lists.gnu.org/archive/html/gnuboot-patches/2025-05/msg00005.html
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>