manual: improve the description of Canoeboot.
The "Because of that it can be harder for users to use." is not
clear. My assumption was that features and computers came and go, but
I didn't manage to confirm that and this slept through the review
process.

What happens instead is that some things are dropped and don't seem to
come back. An example of that is the build of installation utilities
for i945 ThinkPads.

Reported-by: Leah Rowe <leah@libreboot.org>
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>