tests: lint: fix grub.cfg check on Trisquel 11 (aramo).
Running tests/lint on Trisquel 11 results in the following error:
    +---------------------+
    | Running lint tests: |
    +---------------------+
    [...]
    [ OK ] manual/autogen.sh

    In resources/grub/config/grub.cfg line 23:
    set prefix=(memdisk)/boot/grub
               ^-- SC1088 (error): Parsing stopped here.
               Invalid use of parentheses?

    For more information:
      https://www.shellcheck.net/wiki/SC1088 --
      Parsing stopped here. Invalid use...
    [ !! ] resources/grub/config/grub.cfg failed

While on a recent Guix there is no such error.

Since we already have '# shellcheck disable=SC1088', there is no way
to avoid this error on Trisquel 11. Because of that shellcheck was
replaced with grub-script-check which also works on Trisquel 11, but
the big downside is that grub-script-check will not check many things
(for instance it doesn't detect incorrect commands or assignments, and
it even parse correctly a big part of a file produced with 'dmesg').

This was broken from the start on Trisquel 11 (aramo) with the commit
6f6e7315ebf0f28e40dda90055310ece4b873531 ("resources: grub: config:
(ab)use shellcheck to check it.").

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>