coreboot: x200_8mb: remove debugging configurations.
In the commit 5535113f1dace4b99076bf5ac8a300789a7e6d2b ("coreboot:
x200_8mb: add debugging configuration"), debug images were added for
the X200, but Adrien who did the commit, he was not familiar with the
ThinkPad X200, and I didn't spot the problem during the review.

The issue we have is that the ThinkPad X200 has no builtin serial
port, and so even with a dock, we can't get a serial port.

The solution would be to add support for express-card serial ports,
but we don't have such cards to test if this really works, so until we
find somebody that can test that, we remove debug images for the X200.

And having debug images without serial ports or ways to get back the
log reduces a lot the usefulness of such images. Having such image
would also means that we need to differentiate between debug images
with and without serial ports into the documentation, which would then
complexity things for users for very little benefits.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>