manual: move internal SD card readers to new computers subsections.
We want to have computers subsection to be able to document very
specific hardware, like which boot flash variants exist for a given
computer, or which CPUs are available, what is removable or soldered,
what is not, etc.

However internal card readers are currently in a section about
peripherals makes which then makes it complicated to understand where
to add information about hardware support.

So this moves the SD card readers information in computers
subsections. But if these section would only have some information
about microSD readers, they would look a bit strange. So a bit more
information was added along the way such as the computer vendor,
product, max RAM, the CPU available, etc.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>