checkpatch.scm: check if the patch has a valid commit hash.
We want to require this to be able to later on depend on git to be
able to (much more easily and reliably) find out if a file exist at
the commit hash of a certain patch. To do that the patch needs to have
a valid git commit hash.

Checking for file existence can be required to check if translations
were updated, as there is no point in adding a warning if the file
doesn't exist once the commit is applied, and checking current files
with older commit also don't make any sense.

As for the implementation, if we put the check in the function run by
run-end-rules, we can end end up with some confusing output:

    $ ./scripts/checkpatch.scm 0001-no-hash.patch
    [...]
    WARNING: Overview chapter and node are not aligned.

    WARNING: What is GNU Boot section and node are not aligned.

    WARNING: boot software subsection and node are not aligned.

    WARNING: distribution subsection and node are not aligned.

    WARNING: Helping GNU Boot chapter and node are not aligned.

    Error: the patch has no commit hash.

so it needs to run before the other checks.

Running the check inside functions that run for each line of the patch
is also not very efficient or intuitive, so it had to be put inside
the function run by set-defaults (which in turn required to be able to
access parse-results as we don't want to duplicate the parsing code to
know if a valid commit hash was found or not).

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>