packages: grub: don't silently fail downloads.
Before this change, if the GRUB source code was downloaded but not the
gnulib source code, the build system would mistakenly assume that GRUB
was downloaded correctly.

This makes the download atomic in the sense that it either succeed
completely or fails completely.

However having the 'module' task check if the download is done doesn't
really make sense as the information is easier to isolate inside the
download task. In addition, the interface is generic enough to be
adopted by other tasks as well to at the end get a more unified and
robust way to handle dependencies.

Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>
Acked-by: Adrien 'neox' Bourmault <neox@gnu.org>