packages: memtest86plus: don't silently fail downloads.
Before this change, if the memtest86+ source code was downloaded,
however if it was interrupted when applying patches or when changing
git revision, the build system would mistakenly assume that GRUB was
downloaded correctly and that it is now using the right revision.

This makes the download atomic in the sense that it either succeed
completely or fails completely.

However having the 'module' task check if the download is done doesn't
really make sense as the information is easier to isolate inside the
download task. In addition, the interface is generic enough to be
adopted by other tasks as well to at the end get a more unified and
robust way to handle dependencies.
Signed-off-by: Denis 'GNUtoo' Carikli <GNUtoo@cyberdimension.org>